add script log panel, remove annoying log

+++ include/globed/core/RoomManager.hpp
@@ -20,8 +20,8 @@ public:
     int getRoomOwner();
     std::optional<uint8_t> pickServerId();
     RoomSettings& getSettings();
-    void setAttemptedPasscode(uint64_t code);
-    uint64_t getPasscode();
+    void setAttemptedPasscode(uint32_t code);
+    uint32_t getPasscode();
 
     uint16_t getCurrentTeamId();
     std::optional<RoomTeam> getCurrentTeam();
@@ -34,7 +34,7 @@ private:
     ~RoomManager() = default;
 
     uint32_t m_roomId = 0;
-    uint64_t m_passcode = 0;
+    uint32_t m_passcode = 0;
     int m_roomOwner = 0;
     std::string m_roomName;
     RoomSettings m_settings{};
